#550 In-note-search doesn't jump to occurrence of searchstring
Divide et impera
